requirements.txt
python-dotenv==1.0.0
msal==1.25.0
requests==2.31.0 